Remove YAML output options from dsari-info
Mini-rant, probably not suitable for a commit message:

This commit began as adding YAML config loading functionality, as
dsari-info already had YAML dump functionality.  Ultimately it ended up
bloating the code and forcing a lot of assumptions.  (If the main config
is loaded as dsari.yaml, should we look in config.d for *.json and/or
*.yaml?  If the main config is YAML, should dsari-info dump-config
default to YAML?  Etc.)  Usually more choice is a good thing, but this
felt mildly like bloat.

On a personal note, I've never really liked YAML.  It's easy to write,
but that can be a double edged sword, since its type detection is so
loose and not immediately obvious.  In "foo: bar", "bar" is immediately
recognizable as a string, but then you could easily misinterpet "foo:
null".  JSON is pure in its type representation: dicts are {}, lists are
[], strings are double quote encapsulated, everything else is not.

My biggest annoyance with JSON is not supporting superfluous trailing
commas in lists/dicts, and lack of support for single quote
encapsulation.  If it weren't for those, JSON would be a nearly drop-in
replacement for Python representation.

Also, everything else in dsari is JSON.  Extra data is stored in the
database as JSON.  Trigger data is parsed as JSON; would we add support
for parsing YAML triggers?

I'm a little sad removing functionality, but if you're exporting dsari
data using dsari-info into a non-human-readable format (OTOH, tabular is
basically human readable) and do actually want YAML format, there are
trivial ways to convert JSON to YAML on the fly.

Okay, mini rant over.